+++ src/main/java/sopt/makers/authentication/database/rdb/repository/auth/PhoneVerificationRemover.java
@@ -1,0 +1,21 @@
+package sopt.makers.authentication.database.rdb.repository.auth;
+
+import sopt.makers.authentication.domain.auth.PhoneVerification;
+
+import jakarta.transaction.Transactional;
+
+import org.springframework.stereotype.Component;
+
+import lombok.RequiredArgsConstructor;
+
+@Component
+@Transactional
+@RequiredArgsConstructor
+public class PhoneVerificationRemover {
+
+  private final PhoneVerificationJpaRepository jpaRepository;
+
+  public void remove(PhoneVerification phoneVerification) {
+    jpaRepository.deleteByVerification(phoneVerification);
+  }
+}
